Complete header/hero and start card section
When working on the hero portion, the image placeholder's dimensions
may need revised. The use of min-width may be incorrect so further
research will need done on the best approach here. Additionally,
there appears to be some unwanted padding below the hero text on the
left side as my defined gap seems larger between hero text and subtext
compared to subtext and signup button.

Additionally, the card section of the page is beginning to take shape
with some starting flexbox css added. More development on this to come.